:building_construction: New architecture fully implemented
With this new changes, instead of having the core knowing about the UI, this decouples the core from whatever UI we have by exposing queries and mutations through a mediator. Right now the CLI works exactly the same as before but with the new patterns, as seen, instead of having the state initialized in the beginning and modified through calls in the CLI layer, we now retrieve whatever information we need through the `query` function and execute mutations through the `mutate` function.