:recycle: Refactor concert event handling
This introduces huge changes when it comes to event handling in a concert. Instead of manually implementing everything, we now expose an object that can be passed into a common method that takes care of things like checking the number of times an action was performed, applying penalisations for over-performing actions and the actual calculation of the points given for a certain action.