:recycle: Refactor characters to have a single source of truth
This moves all the characters that are created in-game to the `Characters` map inside of the state. All other places (current playable character, band members) are only references to this map.

The only exception is the producer in a studio, which is pending a revision on how we create studio producers. Since right now they're being loaded from a JSON file I'll keep it there, but this should be gone in the future as well.