:recycle: Refactor into situations
Situations are a way to signify the current event that a character is participating in. This way we can distinguish between things like whether the user is exploring the world or in a concert, keeping that information on the state so that they can exit the game mid-concert without losing progress.