:card_file_box: Move world out of state
With this commit savegames will no longer contain the game map and instead it'll be generated the first time it's requested by the game and cached for each successive call. This is done mainly because saving the world in the savegame meant extra spaced taken by the file that was not really necessary (because there was no changing information anyway) and also because any modification to the world corrupted the savegame.