:building_construction: Fans in bands + initial fame gaining
Bands now have the concept of fans instead of fame, which makes it a bit
easier to compute how to gain new fans instead of having to rely on a
percentage for everything. A fame is still estimated based on the band's
genre market to calculate certain things, but fans will be the
"user-facing" metric. This (huge) commit also introduces some initial
fame gaining from the streams that an album receives.