:necktie: Simplify world map
With this commit, the world map is no longer a graph of rooms inside a graph of places and instead is a simple list of places that the user can visit, with all interactions available from the beginning.

In the cases where this doesn't make sense (ie: during concerts), we can use situations to differentiate between different rooms or to simulate context-dependent interactions.