:sparkles: Add calendar to phone
This puts together the concerts and the flights in one app so that it's easier to see the upcoming schedule. Removes the schedule visualizer because it's redundant, but keeps upcoming flights in the Flights app because that one shows all the upcoming ones and not just the ones for the current month.