:recycle: Refactor ticking logic
This addressed a couple of issues and pain points that arose after introducing the time-based attribute change, namely some associated effects should prevent the chain of effects from continuing but there was no way of doing so. This lead to some funny bugs, like getting hospitalized during sleep would make the character travel back in time and lose a bunch of their attributes on the way.

This introduces the concept of `AssociatedEffectType`, which allows the associated effect handlers to tell the ticking logic whether the associated effects should let the process continue (ContinueChain) or discard the rest (BreakChain).

This also removes the time checking on the ticking logic, which will be re-introduced later.