:bug: Fix duplicates when practicing songs
Previously, when practicing songs that were part of an album, the game would re-add that song to the finished songs, essentially duplicating it. This made me re-think the way that albums store that info, which is basically duplicating the songs and it made for this weird way of retrieving songs via the repertoire (now deleted) which had to mash together all the different sources of songs.

With this new model we now use the song repertoire as the single source of truth and albums only reference it via an ID instead of duplicating the songs in the track list.