:bug: Fix `scheduledAroundDate`
Previously it relied on the top of the list containing the latest concert, but since we don't insert it ordered by date it didn't make any sense and it caused issues when multiple concerts were scheduled. Instead the function now goes through all the concerts and chooses the ones that are around the current date.