:bug: Fix crash when creating game
Apparently this has been broken since Moodlets because of the way we apply them. In the first tick of the game (GameStarted) we attempt to apply moodlets on the current character, which of course fails because the state is not yet available. Oopsie! We need some E2E tests over here.