:sparkles: Generate known people in places
Now there's a tiny chance of also encountering someone that the character already knows.

As a fix in this commit the random gen switching has been fixed, apparently before we were never actually switching to the new implementation (was this introduced in .NET 8? no idea), but now it's fixed and with that the need to reset back to the default implementation after each test which is now done with a disposable.