:sparkles: Improve game pace
The previous implementation of the calendar made the game absurdly long. After testing this approach in the Playdate port of the game, I realized that following a Sims-like calendar where there's just seasons with a few days instead makes the game much more entertaining, so switching in the desktop game as well! Now instead of months there's seasons, with each season having 21 days (to have a squared calendar and not have to deal with day offsets)