:loud_sound: Improve logging system
Messages are now saved in-memory while the game is still ongoing
and written to the file when exitting. Also, we don't just keep
appending forever and ever and simply replace the contents of the
file every time that the game is closed. Farewell to those 20MB
text files I've accumulated :^)