:sparkles: Revamp cities with streets, metro system and weather
This chunky boy of a PR finally addresses one of my main pain points with the game: cities were boring! My initial idea was to do the huge work of adding descriptors to each street and then, through many many many layers of matches, create _sort of_ personalized descriptions for each street in a city. This proved not to work quite as much as I imagined, so I turned to attempt using a local LLM to do the work for each description.

There's still a bunch of work to do with this, but as a first draft I think it works pretty well. Major points of the PR are:
- Added a local model that is downloaded on build time, initialized in a separate mailbox and accepts generating any sort of description based on a prompt. For now it's only used to describe places via the `look` command, but I'm planning to expand the usage around the game for dialogues and other descriptions like concert results.
- Streets are back in the game map with a cute little street graph to traverse city zones.
- Added a metro system to travel between zones.
- Added a weather system that follows a realistic-ish model of the city in that season.
- The map command now displays directions instead of teleporting the player there immediately.
- Mexico DC, Tokyo and Sydney have been removed for now since it was already a huge work to adapt the cities that did make the cut. Planning to add them back at some point.
- Added a City Explorer app made with Avalonia that helps visualizing the city graph, metro and the places inside of it.